Clue Vim in on syntax in an autocmd context
As to Vim patch 9.0.1930, the autocommands written in Vim9
syntax and :def-ined in a legacy script fail with E492 at
their events that have been defined with the invocation of
their :def function from _another_ :def function unless
an indirection with a legacy :function have been introduced
(which is but an implementation artifact).  So make do with
:def -> :function -> :def (with autocommands) instead of
more intuitive :def -> :def (with autocommands) whenever
caller functions can be either, Vim9 or legacy.

See https://github.com/vim/vim/issues/13179.